Run did benches once & clean up
