+++ identity/Cargo.toml
@@ -25,9 +25,6 @@ tokio = { version = "1.3", features = ["full"] }
 name = "benchmark"
 harness = false
 
-[[bench]]
-name = "diff_chain"
-harness = false
 
 [features]
 default = ["identifier", "credential", "iota"]
