+++ identity/benches/diff_chain.rs
@@ -1,10 +1,5 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
-use criterion::criterion_group;
-use criterion::criterion_main;
-use criterion::BenchmarkId;
-use criterion::Criterion;
-use criterion::Throughput;
 use identity::iota::MessageId;
 use identity_core::common::Timestamp;
 use identity_core::crypto::KeyPair;
@@ -14,7 +9,7 @@ use identity_iota::did::Document;
 use identity_iota::did::DocumentDiff;
 use identity_iota::tangle::TangleRef;
 
-fn setup() -> (Document, KeyPair) {
+pub fn setup_diff_chain_bench() -> (Document, KeyPair) {
   let keypair: KeyPair = KeyPair::new_ed25519().unwrap();
   let mut document: Document = Document::from_keypair(&keypair).unwrap();
   document.sign(keypair.secret()).unwrap();
@@ -23,7 +18,11 @@ fn setup() -> (Document, KeyPair) {
   (document, keypair)
 }
 
-fn update_diff_chain(n: usize, document: Document, keypair: &KeyPair) {
+pub fn create_diff_chain(document: Document) {
+  let _ = DocumentChain::new(AuthChain::new(document).unwrap());
+}
+/// Creates a diff chain and updates it `n` times
+pub fn update_diff_chain(n: usize, document: Document, keypair: &KeyPair) {
   let mut chain: DocumentChain;
   chain = DocumentChain::new(AuthChain::new(document).unwrap());
 
@@ -41,20 +40,3 @@ fn update_diff_chain(n: usize, document: Document, keypair: &KeyPair) {
     assert!(chain.try_push_diff(diff).is_ok());
   }
 }
-
-fn bench_diff_chain_updates(c: &mut Criterion) {
-  static ITERATIONS: &[usize] = &[10, 100, 200];
-  let (doc, keys) = setup();
-
-  let mut group = c.benchmark_group("update_diff_chain");
-  for size in ITERATIONS.iter() {
-    group.throughput(Throughput::Elements(*size as u64));
-    group.bench_with_input(BenchmarkId::from_parameter(size), size, |b, &size| {
-      b.iter(|| update_diff_chain(size, doc.clone(), &keys));
-    });
-  }
-  group.finish();
-}
-
-criterion_group!(benches, bench_diff_chain_updates);
-criterion_main!(benches);
